requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
importlib-resources==6.4.5
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.8.0
networkx==3.1
numpy
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pkgutil-resolve-name==1.3.10
plotly==5.24.1
protobuf==5.28.1
pyarrow==17.0.0
pydeck==0.9.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
zipp==3.20.2
